Fix ami update script validation
With #34 merged, there are multiple lines of output printed to the screen as a result of running the `jenkins-update-ec2-ami.py` script. Therefore, we need to make sure that we only inspect the last line of output when determine whether or not the script ran completely. This PR addresses that.